Update sign/verify to take the bundle as json
Updated the `sign/signAttestation/verify` library methods to take and
return the bundle file as json to match the format on disk.

With the current interface a caller would be forced to convert the json
to the Bundle object format which seems like an implementation detail. I
ran into this implementing bundle verification in npm where we want to
use the `Sigstore.verify` method to verify the bundle.

Opening up for comments/feedback.

Signed-off-by: Philip Harrison <philip@mailharrison.com>