Add basic help and version command line option
This commit adds very basic command line `help` and `version` command
line options.

The motivation for this that I tried running sigstore --help out of
habit but that returned `Unknown command`. Perhaps others will do this
same and this commit could be a start of adding more options/details.

Perhaps this can be expanded upon and cleaned up by using a module
like yargs, but I was not sure if adding a dependency would be
considered ok, and if there might be preferences of which module to
use in that case.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>