Automate creating changesets for dependabot PRs
Hacked together a action to autogenerate changesets for dependabot PRs. I started out thinking it would be super easy to just invoke the changesets cli from a custom workflow but it's not possible to pass in the necessary args, and realised it's doing a lot more than just taking a summary and semver update type so ended up with writing this https://github.com/feelepxyz/dependabot-changesets 

Not sure it's a great idea yet 😆  Thoughts @bdehamer? 

Signed-off-by: Philip Harrison <philip@mailharrison.com>