Add most verify options to `sigstore verify` subcommand
This PR adds more flags to the sigstore verify subcommand to let users access most of the verify functionality that the sigsign-js library has.

Looks like this was discussed at one point in #477 (comment) and that ideally we'd refactor some of these flags in future, but for now I'm just focusing on getting the functionality working.

Signed-off-by: Ashleigh Carr <ashcorr20@gmail.com>